+++ examples/resources/fmc_access_rule/resource.tf
@@ -48,11 +48,8 @@ resource "fmc_access_rule" "example" {
   ]
   source_port_literals = [
     {
-      type      = "PortLiteral"
-      port      = "80"
-      protocol  = "6"
-      icmp_type = "0"
-      icmp_code = "0"
+      protocol = "6"
+      port     = "80"
     }
   ]
   source_port_objects = [
@@ -62,11 +59,9 @@ resource "fmc_access_rule" "example" {
   ]
   destination_port_literals = [
     {
-      type      = "PortLiteral"
-      port      = "80"
-      protocol  = "6"
-      icmp_type = "0"
-      icmp_code = "0"
+      type     = "PortLiteral"
+      port     = "80"
+      protocol = "6"
     }
   ]
   destination_port_objects = [