+++ examples/resources/fmc_access_rules/resource.tf
@@ -50,11 +50,8 @@ resource "fmc_access_rules" "example" {
       ]
       source_port_literals = [
         {
-          type      = "PortLiteral"
-          port      = "80"
-          protocol  = "6"
-          icmp_type = "0"
-          icmp_code = "0"
+          protocol = "6"
+          port     = "80"
         }
       ]
       source_port_objects = [
@@ -64,11 +61,9 @@ resource "fmc_access_rules" "example" {
       ]
       destination_port_literals = [
         {
-          type      = "PortLiteral"
-          port      = "80"
-          protocol  = "6"
-          icmp_type = "0"
-          icmp_code = "0"
+          type     = "PortLiteral"
+          port     = "80"
+          protocol = "6"
         }
       ]
       destination_port_objects = [