+++ internal/provider/resource_fmc_extended_access_list.go
@@ -252,31 +252,19 @@ func (r *ExtendedAccessListResource) Schema(ctx context.Context, req resource.Sc
 							},
 						},
 						"source_port_literals": schema.SetNestedAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Set of objects that represent destination port of traffic (literally specified).").String,
+							MarkdownDescription: helpers.NewAttributeDescription("Set of objects that represent source port of traffic (literally specified).").String,
 							Optional:            true,
 							NestedObject: schema.NestedAttributeObject{
 								Attributes: map[string]schema.Attribute{
-									"type": schema.StringAttribute{
-										MarkdownDescription: helpers.NewAttributeDescription("Type of the object.").AddStringEnumDescription("PortLiteral", "ICMPv4PortLiteral").String,
+									"protocol": schema.StringAttribute{
+										MarkdownDescription: helpers.NewAttributeDescription("IANA protocol number.").AddStringEnumDescription("6", "17").String,
 										Required:            true,
 										Validators: []validator.String{
-											stringvalidator.OneOf("PortLiteral", "ICMPv4PortLiteral"),
+											stringvalidator.OneOf("6", "17"),
 										},
 									},
 									"port": schema.StringAttribute{
 										MarkdownDescription: helpers.NewAttributeDescription("Port number.").String,
-										Optional:            true,
-									},
-									"protocol": schema.StringAttribute{
-										MarkdownDescription: helpers.NewAttributeDescription("IANA port number.").String,
-										Required:            true,
-									},
-									"icmp_type": schema.StringAttribute{
-										MarkdownDescription: helpers.NewAttributeDescription("ICMP type.").String,
-										Optional:            true,
-									},
-									"icmp_code": schema.StringAttribute{
-										MarkdownDescription: helpers.NewAttributeDescription("ICMP code").String,
 										Optional:            true,
 									},
 								},